+++ adminer/drivers/sqlite.inc.php
@@ -133,6 +133,13 @@ if (isset($_GET["sqlite"])) {
 			return array_keys($this->types[0]);
 		}
 
+		function engines(): array {
+			return (min_version(3.37)
+				? array("STRICT", "STRICT, WITHOUT ROWID", "WITHOUT ROWID")
+				: (min_version("3.8.2") ? array("WITHOUT ROWID") : array())
+			);
+		}
+
 		function insertUpdate(string $table, array $rows, array $primary) {
 			$values = array();
 			foreach ($rows as $set) {
@@ -209,7 +216,15 @@ if (isset($_GET["sqlite"])) {
 
 	function table_status($name = "") {
 		$return = array();
-		foreach (get_rows("SELECT name AS Name, type AS Engine, 'rowid' AS Oid, '' AS Auto_increment FROM sqlite_master WHERE type IN ('table', 'view') " . ($name != "" ? "AND name = " . q($name) : "ORDER BY name")) as $row) {
+		foreach (get_rows("SELECT name AS Name, type AS Engine, sql, 'rowid' AS Oid, '' AS Auto_increment FROM sqlite_master WHERE type IN ('table', 'view') " . ($name != "" ? "AND name = " . q($name) : "ORDER BY name")) as $row) {
+			if ($row["Engine"] == "table") {
+				$suffix = preg_replace('~.*\)~s', '', $row["sql"]); // table options are after the last parenthesis
+				$row["Engine"] = implode(", ", array_filter(array(
+					(preg_match('~\bSTRICT\b~i', $suffix) ? "STRICT" : ""),
+					(preg_match('~\bWITHOUT\s+ROWID\b~i', $suffix) ? "WITHOUT ROWID" : ""),
+				)));
+			}
+			unset($row["sql"]);
 			$row["Rows"] = get_val("SELECT COUNT(*) FROM " . idf_escape($row["Name"]));
 			$return[$row["Name"]] = $row;
 		}
@@ -400,7 +415,7 @@ if (isset($_GET["sqlite"])) {
 	}
 
 	function alter_table($table, $name, $fields, $foreign, $comment, $engine, $collation, $auto_increment, $partitioning) {
-		$use_all_fields = ($table == "" || $foreign);
+		$use_all_fields = ($table == "" || $foreign || $engine !== null);
 		foreach ($fields as $field) {
 			if ($field[0] != "" || !$field[1] || $field[2]) {
 				$use_all_fields = true;
@@ -426,7 +441,7 @@ if (isset($_GET["sqlite"])) {
 			if ($table != $name && !queries("ALTER TABLE " . table($table) . " RENAME TO " . table($name))) {
 				return false;
 			}
-		} elseif (!recreate_table($table, $name, $alter, $originals, $foreign, $auto_increment)) {
+		} elseif (!recreate_table($table, $name, $alter, $originals, $foreign, $auto_increment, array(), "", "", $engine)) {
 			return false;
 		}
 		if ($auto_increment) {
@@ -451,8 +466,7 @@ if (isset($_GET["sqlite"])) {
 	* @param string $drop_check CHECK constraint to drop
 	* @param string $add_check CHECK constraint to add
 	*/
-	function recreate_table(string $table, string $name, array $fields, array $originals, array $foreign, string $auto_increment = "", $indexes = array(), string $drop_check = "", string $add_check = ""): bool {
-		$suffix = "";
+	function recreate_table(string $table, string $name, array $fields, array $originals, array $foreign, string $auto_increment = "", $indexes = array(), string $drop_check = "", string $add_check = "", ?string $engine = null): bool {
 		if ($table != "") {
 			if (!$fields) {
 				foreach (fields($table) as $key => $field) {
@@ -507,19 +521,6 @@ if (isset($_GET["sqlite"])) {
 					$foreign[] = " " . format_foreign_key($foreign_key);
 				}
 			}
-			$options = array();
-			if (min_version(3.37)) {
-				$row = first(get_rows("PRAGMA table_list(" . table($table) . ")"));
-				if ($row["wr"]) {
-					$options[] = "WITHOUT ROWID";
-				}
-				if ($row["strict"]) {
-					$options[] = "STRICT";
-				}
-			} elseif (preg_match('~\)\s*WITHOUT\s+ROWID\s*$~i', get_val("SELECT sql FROM sqlite_master WHERE type = 'table' AND name = " . q($table)))) {
-				$options[] = "WITHOUT ROWID";
-			}
-			$suffix = ($options ? " " . implode(", ", $options) : "");
 			queries("BEGIN");
 		}
 		$changes = array();
@@ -539,7 +540,10 @@ if (isset($_GET["sqlite"])) {
 			$changes[] = "  CHECK ($add_check)";
 		}
 		$temp_name = ($table == $name ? "adminer_$name" : $name);
-		if (!queries("CREATE TABLE " . table($temp_name) . " (\n" . implode(",\n", $changes) . "\n)$suffix")) {
+		if ($engine === null && $table != "") {
+			$engine = idx(table_status1($table), "Engine");
+		}
+		if (!queries("CREATE TABLE " . table($temp_name) . " (\n" . implode(",\n", $changes) . "\n)" . (in_array($engine, driver()->engines()) ? " $engine" : ""))) {
 			// implicit ROLLBACK to not overwrite connection()->error
 			return false;
 		}